feat: moving random function to index file
